Introduce a shared definition for archive contents
The contents of different camp archives wered mixed up, especially '24
just mostly consisted of '23 contents. Probably copy-paste error.

To make it easier to create archives going forward, this introduces a
type to clarify which contents are expected on archive page, and allows
to just pass in the content fragments already defined for the actual
camp page.

Uses this shared definition for '24 archive page for now. Should also
use this for '23 and '25 going forward.